third_party/WebKit/LayoutTests/platform/win/http/tests/security/secureContexts/unauthenticated_sandbox-expected.txt

CONSOLE ERROR: line 16: Content Security Policies delivered via a <meta> element may not contain the sandbox directive.
This is a testharness.js-based test.
PASS unauthenticated origin is insecure. 
PASS Frames inside sandboxed frames are insecure 
Harness: the test ran to completion.
